+++ entrypoints/data_maintainer.py
@@ -1,15 +1,22 @@
 """Model for generating data for training our models"""
 
 import click
+from sqlalchemy.orm import Session
 
+from freekick import _logger
+from freekick.datastore import DEFAULT_ENGINE
+from freekick.datastore.repository import SQLAlchemyRepository
 from freekick.datastore.util import (
+    CSVUtils,
     DataScraper,
     DataStore,
+    DBUtils,
     League,
     get_league_data_container,
 )
 
 UPDATE_TYPES = ["team_rating", "player_rating", "match", "current_season"]
+_logger.setLevel("INFO")
 
 
 @click.group()
@@ -48,11 +55,11 @@ def update(data_type, league, persist):
         # read_stitch_raw_data(league=League[league], persist=persist)
         # league_container.read_stitch_raw_data(league=league, persist=persist)
     elif data_type == "current_season":
-        # TODO: Also update for DB
-        league_data = league_container(
-            datastore=DataStore.CSV, repository=None
-        )
-        league_data.update_current_season(persist=persist)
+        session = Session(DEFAULT_ENGINE)
+        repo = SQLAlchemyRepository(session)
+        for store in DataStore:
+            league_data = league_container(datastore=store, repository=repo)
+            league_data.update_current_season(persist=persist)
 
 
 @click.command()
@@ -64,8 +71,23 @@ def league(list):
         print(League._member_names_)
 
 
+@click.command()
+@click.argument("name")
+@click.argument("code")
+@click.argument("league")
+def add_team(name, code, league):
+    teams = [
+        DBUtils.new_team(team_code=code, team_name=name, league=League[league])
+    ]
+    session = Session(DEFAULT_ENGINE)
+    repo = SQLAlchemyRepository(session)
+    DBUtils.add_teams(teams=teams, repository=repo)
+    CSVUtils.add_teams(teams=teams)
+
+
 cli.add_command(update)
 cli.add_command(league)
+cli.add_command(add_team)
 
 if __name__ == "__main__":
     cli()