+++ CHANGELOG.rst
@@ -10,8 +10,8 @@ Version 0.2
 
   - _LEAGUES has been renamed to League and is now an Enum.
   - Migrate handlers from Flask Blueprint to flask-restx
-  - DecisionTreeClassifier is not the default estimator instead of previous LogisticRegression.
-  - Predictions now support specifying match date and time.
+  - DecisionTreeClassifier is now the default estimator instead of the previous LogisticRegression.
+  - Predictions now support specifying match date, time and attendance from UI.
   - Added swagger endpoint at '/swagger'
   - Stratify training dataset